lock

+++ out/magic/emulate.js
@@ -9,14 +9,18 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
 };
 Object.defineProperty(exports, "__esModule", { value: true });
 const db_1 = require("../uq-api/db");
+const gfuncs_1 = require("../gfuncs");
 const const_1 = require("../const");
 const GroupSize = 30;
 const MaxGroup = 80;
 function emulateAtDay(date) {
     return __awaiter(this, void 0, void 0, function* () {
-        let runner = yield db_1.getRunner('mi');
-        let em = new EmulateMagic(runner);
+        if (gfuncs_1.RemoteIsRun())
+            return;
+        gfuncs_1.RemoteRun(true);
         try {
+            let runner = yield db_1.getRunner('mi');
+            let em = new EmulateMagic(runner);
             console.log('emulate begin day: ' + date);
             let year = Math.floor(date / 10000);
             let month = date % 10000;
@@ -34,14 +38,18 @@ function emulateAtDay(date) {
         catch (err) {
             console.log(err);
         }
+        gfuncs_1.RemoteRun(false);
     });
 }
 exports.emulateAtDay = emulateAtDay;
 function emulateAll() {
     return __awaiter(this, void 0, void 0, function* () {
-        let runner = yield db_1.getRunner('mi');
-        let em = new EmulateMagic(runner);
+        if (gfuncs_1.RemoteIsRun())
+            return;
+        gfuncs_1.RemoteRun(true);
         try {
+            let runner = yield db_1.getRunner('mi');
+            let em = new EmulateMagic(runner);
             let sql = 'delete from tv_神奇公式模拟结果 where 1=1';
             yield runner.sql(sql, []);
             for (let yearlen = 5; yearlen >= 1; --yearlen) {
@@ -60,55 +68,63 @@ function emulateAll() {
         catch (err) {
             console.log(err);
         }
+        gfuncs_1.RemoteRun(false);
     });
 }
 exports.emulateAll = emulateAll;
 function allStocksAvg(begin, end) {
     return __awaiter(this, void 0, void 0, function* () {
-        let runner = yield db_1.getRunner('mi');
-        let ret = [];
-        let pageStart = 0, pageSize = 500;
-        for (;;) {
-            let ids = yield runner.tuidSeach('股票', const_1.DefaultUnit, undefined, undefined, '', pageStart, pageSize);
-            let arr = ids[0];
-            if (arr.length > pageSize) {
-                let top = arr.pop();
-                ret.push(...arr);
-                pageStart = arr[pageSize - 1].id;
+        if (gfuncs_1.RemoteIsRun())
+            return;
+        gfuncs_1.RemoteRun(true);
+        try {
+            let runner = yield db_1.getRunner('mi');
+            let ret = [];
+            let pageStart = 0, pageSize = 500;
+            for (;;) {
+                let ids = yield runner.tuidSeach('股票', const_1.DefaultUnit, undefined, undefined, '', pageStart, pageSize);
+                let arr = ids[0];
+                if (arr.length > pageSize) {
+                    let top = arr.pop();
+                    ret.push(...arr);
+                    pageStart = arr[pageSize - 1].id;
+                }
+                else {
+                    ret.push(...arr);
+                    break;
+                }
             }
-            else {
-                ret.push(...arr);
-                break;
-            }
-        }
-        let count = ret.length;
-        let rCount = 0;
-        let sum = 0;
-        let dayBegin = begin > 0 ? begin : 20110101;
-        let dayEnd = end > 0 ? end : 20190101;
-        for (let i = 0; i < count; ++i) {
-            let code = ret[i];
-            let { id } = code;
-            let pret = yield runner.query('getStockRestorePrice', const_1.DefaultUnit, undefined, [id, dayBegin, dayEnd]);
-            let parr = pret;
-            let r = parr[0];
-            if (r !== undefined) {
-                let { priceBegin, priceEx, bonus, bday } = r;
-                if (bday - dayBegin < 300) {
-                    priceEx = priceEx + bonus;
-                    if (priceBegin > 0 && priceEx > 0) {
-                        ++rCount;
-                        let one = (priceEx / priceBegin - 1) * 100;
-                        sum += one;
+            let count = ret.length;
+            let rCount = 0;
+            let sum = 0;
+            let dayBegin = begin > 0 ? begin : 20110101;
+            let dayEnd = end > 0 ? end : 20190101;
+            for (let i = 0; i < count; ++i) {
+                let code = ret[i];
+                let { id } = code;
+                let pret = yield runner.query('getStockRestorePrice', const_1.DefaultUnit, undefined, [id, dayBegin, dayEnd]);
+                let parr = pret;
+                let r = parr[0];
+                if (r !== undefined) {
+                    let { priceBegin, priceEx, bonus, bday } = r;
+                    if (bday - dayBegin < 300) {
+                        priceEx = priceEx + bonus;
+                        if (priceBegin > 0 && priceEx > 0) {
+                            ++rCount;
+                            let one = (priceEx / priceBegin - 1) * 100;
+                            sum += one;
+                        }
                     }
                 }
             }
+            if (rCount > 0) {
+                sum = sum / rCount;
+                console.log('股数: ' + rCount + '  平均涨幅：' + sum + ' dayBegin=' + dayBegin + ' dayEnd=' + dayEnd);
+                yield runner.mapSave('股市平均涨幅', const_1.DefaultUnit, undefined, [dayBegin, dayEnd, sum, rCount]);
+            }
         }
-        if (rCount > 0) {
-            sum = sum / rCount;
-            console.log('股数: ' + rCount + '  平均涨幅：' + sum + ' dayBegin=' + dayBegin + ' dayEnd=' + dayEnd);
-            yield runner.mapSave('股市平均涨幅', const_1.DefaultUnit, undefined, [dayBegin, dayEnd, sum, rCount]);
-        }
+        catch (err) { }
+        gfuncs_1.RemoteRun(false);
     });
 }
 exports.allStocksAvg = allStocksAvg;
