+++ src/magic/emulate.ts
@@ -1,14 +1,17 @@
 import { getRunner, Runner } from '../uq-api/db';
-import { sleep, checkToDateInt, checkNumberNaNToZero } from '../gfuncs';
+import { sleep, checkToDateInt, checkNumberNaNToZero, RemoteIsRun, RemoteRun } from '../gfuncs';
 import { DefaultUnit } from '../const';
 
 const GroupSize = 30;
 const MaxGroup = 80;
 
 export async function emulateAtDay(date: number) {
-  let runner: Runner = await getRunner('mi');
-  let em = new EmulateMagic(runner);
+  if (RemoteIsRun())
+    return;
+  RemoteRun(true);
   try {
+    let runner: Runner = await getRunner('mi');
+    let em = new EmulateMagic(runner);
     console.log('emulate begin day: ' + date);
     let year = Math.floor(date / 10000);
     let month = date % 10000;
@@ -26,13 +29,16 @@ export async function emulateAtDay(date: number) {
   catch (err) {
     console.log(err);
   }
+  RemoteRun(false);
 }
 
 export async function emulateAll() {
-  let runner: Runner = await getRunner('mi');
-
-  let em = new EmulateMagic(runner);
+  if (RemoteIsRun())
+    return;
+  RemoteRun(true);
   try {
+    let runner: Runner = await getRunner('mi');
+    let em = new EmulateMagic(runner);
     let sql = 'delete from tv_神奇公式模拟结果 where 1=1';
     await runner.sql(sql, []);
     for (let yearlen = 5; yearlen >= 1; --yearlen) {
@@ -43,7 +49,7 @@ export async function emulateAll() {
             break;
           let p = { year: year, month: month, day: 1, date: date, yearlen: yearlen };
           await em.proceeOneDay(p);
-          console.log('emulate end. yearlen: ' +yearlen + '  day: ' + date);
+          console.log('emulate end. yearlen: ' + yearlen + '  day: ' + date);
         }
       }
     }
@@ -51,57 +57,65 @@ export async function emulateAll() {
   catch (err) {
     console.log(err);
   }
+  RemoteRun(false);
 }
 
 export async function allStocksAvg(begin: number, end: number) {
-  let runner: Runner = await getRunner('mi');
+  if (RemoteIsRun())
+    return;
+  RemoteRun(true);
+  try {
+    let runner: Runner = await getRunner('mi');
 
-  let ret: any[] = [];
-  let pageStart = 0, pageSize = 500;
-  for (; ;) {
-    let ids = await runner.tuidSeach('股票', DefaultUnit, undefined, undefined, '', pageStart, pageSize);
-    let arr = ids[0];
-    if (arr.length > pageSize) {
-      let top = arr.pop();
-      ret.push(...arr);
-      pageStart = arr[pageSize - 1].id;
+    let ret: any[] = [];
+    let pageStart = 0, pageSize = 500;
+    for (; ;) {
+      let ids = await runner.tuidSeach('股票', DefaultUnit, undefined, undefined, '', pageStart, pageSize);
+      let arr = ids[0];
+      if (arr.length > pageSize) {
+        let top = arr.pop();
+        ret.push(...arr);
+        pageStart = arr[pageSize - 1].id;
+      }
+      else {
+        ret.push(...arr);
+        break;
+      }
     }
-    else {
-      ret.push(...arr);
-      break;
-    }
-  }
-  let count = ret.length;
+    let count = ret.length;
 
-  let rCount = 0;
-  let sum = 0;
-  let dayBegin = begin > 0 ? begin : 20110101;
-  let dayEnd = end > 0 ? end : 20190101;
-  for (let i = 0; i < count; ++i) {
-    let code = ret[i];
-    let { id } = code;
-    let pret = await runner.query('getStockRestorePrice', DefaultUnit, undefined, [id, dayBegin, dayEnd]);
-    let parr = pret as any[];
-    let r = parr[0];
-    if (r !== undefined) {
-      let { priceBegin, priceEx, bonus, bday } = r as { priceBegin: number, priceEx: number, bonus: number, bday: number, eday: number };
-      if (bday - dayBegin < 300) {
-        priceEx = priceEx + bonus;
-        if (priceBegin > 0 && priceEx > 0) {
-          ++rCount;
-          let one = (priceEx / priceBegin - 1) * 100;
-          sum += one;
+    let rCount = 0;
+    let sum = 0;
+    let dayBegin = begin > 0 ? begin : 20110101;
+    let dayEnd = end > 0 ? end : 20190101;
+    for (let i = 0; i < count; ++i) {
+      let code = ret[i];
+      let { id } = code;
+      let pret = await runner.query('getStockRestorePrice', DefaultUnit, undefined, [id, dayBegin, dayEnd]);
+      let parr = pret as any[];
+      let r = parr[0];
+      if (r !== undefined) {
+        let { priceBegin, priceEx, bonus, bday } = r as { priceBegin: number, priceEx: number, bonus: number, bday: number, eday: number };
+        if (bday - dayBegin < 300) {
+          priceEx = priceEx + bonus;
+          if (priceBegin > 0 && priceEx > 0) {
+            ++rCount;
+            let one = (priceEx / priceBegin - 1) * 100;
+            sum += one;
+          }
         }
       }
     }
-  }
 
-  if (rCount > 0) {
-    sum = sum / rCount;
-    console.log('股数: ' + rCount + '  平均涨幅：' + sum + ' dayBegin=' + dayBegin + ' dayEnd=' + dayEnd);
-    await runner.mapSave('股市平均涨幅', DefaultUnit, undefined,
-      [dayBegin, dayEnd, sum, rCount]);
+    if (rCount > 0) {
+      sum = sum / rCount;
+      console.log('股数: ' + rCount + '  平均涨幅：' + sum + ' dayBegin=' + dayBegin + ' dayEnd=' + dayEnd);
+      await runner.mapSave('股市平均涨幅', DefaultUnit, undefined,
+        [dayBegin, dayEnd, sum, rCount]);
+    }
   }
+  catch (err) { }
+  RemoteRun(false);
 }
 
 class EmulateMagic {
